move dnd delegate
