+++ src/shapes/IText/DraggableTextDelegate.ts
@@ -1,13 +1,17 @@
-import type { Canvas } from '../canvas/canvas_events';
-import { getEnv } from '../env';
-import { DragEventData, DropEventData, TPointerEvent } from '../EventTypeDefs';
-import { Point } from '../point.class';
-import type { IText } from '../shapes/IText/IText';
-import { setStyle } from '../util/dom_style';
-import { cloneDeep } from '../util/internals/cloneDeep';
-import { createCanvasElement } from '../util/misc/dom';
-import { isIdentityMatrix } from '../util/misc/matrix';
-import { TextStyleDeclaration } from '../shapes/Text/StyledText';
+import type { Canvas } from '../../canvas/canvas_events';
+import { getEnv } from '../../env';
+import {
+  DragEventData,
+  DropEventData,
+  TPointerEvent,
+} from '../../EventTypeDefs';
+import { Point } from '../../point.class';
+import type { IText } from './IText';
+import { setStyle } from '../../util/dom_style';
+import { cloneDeep } from '../../util/internals/cloneDeep';
+import { createCanvasElement } from '../../util/misc/dom';
+import { isIdentityMatrix } from '../../util/misc/matrix';
+import { TextStyleDeclaration } from '../Text/StyledText';
 
 /**
  * #### Dragging IText/Textbox Lifecycle
